schema_builder/requirements.txt
fastapi
uvicorn[standard]
aiohttp
beautifulsoup4
python-dotenv
openai
httpx
aiofiles
aiodns
async-timeout
yarl
python-multipart
python-slugify